+++ documentation/Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [unreleased]
 
 ### Changed
+* `setCoupons` and `setShops` on project not index
+* implement `Equatable` and `Hashable` on project
 
 ### Added